pyproject.toml
[project]
name = "svg-text2path"
version = "0.2.0"
description = "Convert SVG text elements to path outlines with HarfBuzz shaping"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Emasoft", email = "713559+Emasoft@users.noreply.github.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics :: Graphics Conversion",
    "Topic :: Text Processing :: Markup :: XML",
]
keywords = ["svg", "text", "path", "convert", "font", "harfbuzz", "vector"]

dependencies = [
    "defusedxml>=0.7.1",
    "fonttools>=4.60.1",
    "numpy>=2.2.6",
    "pillow>=12.0.0",
    "python-bidi>=0.6.7",
    "svg-path>=7.0",
    "uharfbuzz>=0.52.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
lxml = ["lxml>=5.0"]
bs4 = ["beautifulsoup4>=4.12"]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[project.scripts]
text2path = "svg_text2path.cli.main:main"
t2p_convert = "text2path.main:main"
t2p_compare = "text2path.frame_comparer:main"
t2p_font_report = "text2path.font_report:main"
t2p_font_report_html = "text2path.font_report_html:main"
t2p_analyze_path = "text2path.analyze_path:main"
t2p_text_flow_test = "text2path.text_flow_tester:main"

[project.urls]
Homepage = "https://github.com/Emasoft/svg-text2path"
Documentation = "https://github.com/Emasoft/svg-text2path#readme"
Repository = "https://github.com/Emasoft/svg-text2path.git"
Issues = "https://github.com/Emasoft/svg-text2path/issues"
Changelog = "https://github.com/Emasoft/svg-text2path/blob/main/CHANGELOG.md"

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["svg_text2path*", "text2path*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "C4", "SIM"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=4.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]
